refactor(core/query-builder): extract WhereGroupBuilder and shared operator/validator types (#405)
Pure move; WhereOperator/RowValidator/ArrayValidator live in query-builder/types.ts, all re-exported from SelectQueryBuilder.ts.
